(UPDATE) Code cleanup, added inline documentation (CenturyDropDown.js, CheckBox.js, DateDropdown.js, Dialog.js, DropDown.js <-- still needs some cleaning
