+++ src/components/CenturyDropDown.js
@@ -1,12 +1,19 @@
+////////////////////////////////////////////////////////////////////////////////////////
+// CENTURYDROPDOWN.JS /////////////////////////////////////////////////////////////////
+// This component is a selection dropdown to decide the century of the date format ///
+// In the case that the user selects a format with YY instead of YYYY ///////////////
+////////////////////////////////////////////////////////////////////////////////////
+
 import React from "react";
 import { connect } from "react-redux";
 
 // CSS & Styling
 import "semantic-ui-react";
 
-// Action Creators
+// REDUX
 import { century } from "../actions/"
-///////////////////////////////////////////
+
+////////////////////////////////////////////////////
 ///////////////////////////////////////////////////
 
 class CenturyDropDown extends React.Component {
