+++ src/components/CheckBox.js
@@ -1,13 +1,23 @@
+///////////////////////////////////////////////////////////////////////////////////////
+// CHECKBOX.JS ///////////////////////////////////////////////////////////////////////
+// This component displays  a checkbox on the left of each fieldCard ////////////////
+// Giving the user to decide if they want to use that fieldCard in the map or not //
+///////////////////////////////////////////////////////////////////////////////////
+
 import React, { Component } from 'react';
 
 // CSS & Styling
 import { Checkbox } from 'semantic-ui-react';
 import './App.css';
-///////////////////////////////////////////
+
+/////////////////////////////////////////////////
 ////////////////////////////////////////////////
 
 export default class CheckboxExample extends Component {
+
     state = { checked: this.props.isChecked }
+
+    // function that goes back to the fieldCard and changes the color of the fieldCard (previously the cards were green, hints the name)
     toggle = () => {
         this.props.greenCallback()
         this.setState({ checked: !this.props.isChecked })
