+++ src/components/DateDropdown.js
@@ -1,3 +1,9 @@
+///////////////////////////////////////////////////////////////////////////////////////
+// DATEDROPDOWN.JS ///////////////////////////////////////////////////////////////////
+// This component displays  a checkbox on the left of each fieldCard ////////////////
+// Giving the user to decide if they want to use that fieldCard in the map or not //
+///////////////////////////////////////////////////////////////////////////////////
+
 import React from "react";
 import { connect } from "react-redux";
 
@@ -6,32 +12,30 @@ import "semantic-ui-react";
 
 // Action Creators
 import { formatDate } from "../actions/"
-//////////////////////////////////////////
+
+////////////////////////////////////////////////
 ///////////////////////////////////////////////
 
 
 class DateDropdown extends React.Component {
 
-    constructor(props) {
-        super(props);
-        this.state = {
-            currentChosen: null
-        };
-    }
+    state = {
+        currentChosen: null
+    };
 
     // uses the clicked list-item in the dropdown to create an object to be passed into the dropdownUpdate action
     // updates specific object in the redux store
     updateValue = e => {
         const newValue = e.target.value
-        let it = false
+        let needsCenturySelection = false
 
         // Vets the dropdown selection of have 2 or 4 whys for the purposes of using/not using the century dropdown
         if (newValue.includes("YY") && !newValue.includes("YYYY"))
-            it = true
+            needsCenturySelection = true
 
         const obj = {
             dateFormat: newValue,
-            hasTwoYs: it
+            hasTwoYs: needsCenturySelection
         }
         this.props.formatDate(obj)
     }
@@ -74,9 +78,7 @@ class DateDropdown extends React.Component {
 const mapStateToProps = (state) => {
     return {
         hasChosenDropdown: state.hasChosenDropdownOption,
-        hasChosenDateFormat: state.hasChosenDateFormat,
-        hasChosenCentury: state.hasChosenCentury,
-        hasTwoYs: state.hasTwoYs
+        hasChosenDateFormat: state.hasChosenDateFormat
     };
 };
