+++ src/components/Dialog.js
@@ -1,5 +1,15 @@
+////////////////////////////////////////////////////////////////////////////////////////
+// DIALOG.JS //////////////////////////////////////////////////////////////////////////
+// This component displays a preview of your map as you are creating it //////////////
+// When "preview map" in the toolbar is clicked, a small window opens with content //
+////////////////////////////////////////////////////////////////////////////////////
+
 import React, { Component } from 'react';
 
+/////////////////////////////////////////////////
+////////////////////////////////////////////////
+
+// CSS & Styling
 let dialogStyles = {
     width: '800px',
     maxWidth: '100%',
