+++ src/components/sesarOptions.js
@@ -271,4 +271,20 @@ let options = [
 
 ]
 
-module.exports = { options };
+const dateFormatOption = [
+    { title: "Select Date Format" },
+    { title: "DD/MM/YY or DD-MM-YY", value: "substring", type: "date" },
+    { title: "MM/DD/YY or MM-DD-YY", value: "substring", type: "date" },
+    { title: "YY/DD/MM or YY-DD-MM", value: "substring", type: "date" },
+    { title: "YY/MM/DD or YY-MM-DD", value: "substring", type: "date" },
+    { title: "DD/MM/YYYY or DD-MM-YYYY", value: "substring", type: "date" },
+    { title: "MM/DD/YYYY or MM-DD-YYYY", value: "substring", type: "date" },
+    { title: "YYYY/DD/MM or YYYY-DD-MM", value: "substring", type: "date" },
+    { title: "YYYY/MM/DD or YYYY-MM-DD", value: "substring", type: "date" },
+    { title: "MMDDYYYY", value: "substring", type: "date" },
+    { title: "DDMMYYYY", value: "substring", type: "date" },
+    { title: "YYYYDDMM", value: "substring", type: "date" },
+    { title: "YYYYMMDD", value: "substring", type: "date" }
+]
+
+module.exports = { options, dateFormatOption };
